feat(prompts): implement prompts management with categories and favorite features
