+++ lib/core/constants/categories.dart
@@ -1,0 +1,15 @@
+import 'package:khtn_ai_final_project/domain/entities/category.dart';
+
+final List<Category> categories = [
+  Category(name: "Coding", iconPath: "assets/icons/ic_coding.svg"),
+  Category(name: "Career", iconPath: "assets/icons/ic_career.svg"),
+  Category(name: "Business", iconPath: "assets/icons/ic_business.svg"),
+  Category(name: "Education", iconPath: "assets/icons/ic_education.svg"),
+  Category(name: "Marketing", iconPath: "assets/icons/ic_marketing.svg"),
+  Category(name: "Writing", iconPath: "assets/icons/ic_writing.svg"),
+  Category(name: "Fun", iconPath: "assets/icons/ic_fun.svg"),
+  Category(name: "Chatbot", iconPath: "assets/icons/ic_bot.svg"),
+  Category(name: "Productivity", iconPath: "assets/icons/ic_productivity.svg"),
+  Category(name: "CEO", iconPath: "assets/icons/ic_ceo.svg"),
+  Category(name: "Other", iconPath: "assets/icons/ic_other.svg"),
+];
