+++ lib/core/constants/sample_prompts.dart
@@ -1,0 +1,62 @@
+import 'package:khtn_ai_final_project/domain/entities/prompt.dart';
+
+final List<Prompt> samplePrompts = [
+  Prompt(
+    id: "p1",
+    createdAt: "2025-10-20T10:00:00Z",
+    updatedAt: "2025-10-21T09:00:00Z",
+    category: "Productivity",
+    content: "Write a daily plan to maximize focus and minimize distractions.",
+    description: "A prompt to help organize daily priorities effectively.",
+    isPublic: true,
+    language: "en",
+    title: "Daily Focus Planner",
+    userId: "u101",
+    userName: "Alice",
+    isFavorite: true,
+  ),
+  Prompt(
+    id: "p2",
+    createdAt: "2025-10-19T14:30:00Z",
+    updatedAt: "2025-10-20T15:00:00Z",
+    category: "Coding",
+    content:
+        "Explain the difference between stateful and stateless widgets in Flutter.",
+    description:
+        "A coding-related educational question about Flutter UI structure.",
+    isPublic: true,
+    language: "en",
+    title: "Flutter Widget Types",
+    userId: "u102",
+    userName: "Bob",
+    isFavorite: false,
+  ),
+  Prompt(
+    id: "p3",
+    createdAt: "2025-10-18T08:15:00Z",
+    updatedAt: "2025-10-18T09:30:00Z",
+    category: "AI Writing",
+    content: "Generate a short story about an AI that learns to dream.",
+    description: "A creative writing prompt exploring AI consciousness.",
+    isPublic: true,
+    language: "en",
+    title: "The Dreaming AI",
+    userId: "u103",
+    userName: "Charlie",
+    isFavorite: true,
+  ),
+  Prompt(
+    id: "p4",
+    createdAt: "2025-10-17T12:00:00Z",
+    updatedAt: "2025-10-18T08:00:00Z",
+    category: "Career",
+    content: "Give me 5 strategies to improve communication in a remote team.",
+    description: "Useful for professionals managing distributed teams.",
+    isPublic: false,
+    language: "en",
+    title: "Remote Team Communication Tips",
+    userId: "u104",
+    userName: "Diana",
+    isFavorite: false,
+  ),
+];
