+++ lib/domain/entities/category.dart
@@ -1,0 +1,6 @@
+class Category {
+  final String name;
+  final String iconPath;
+
+  Category({required this.name, required this.iconPath});
+}
