+++ lib/presentation/views/prompts/widgets/category_item.dart
@@ -1,0 +1,66 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:khtn_ai_final_project/theme/app_radius.dart';
+
+class CategoryItem extends StatelessWidget {
+  final String categoryName;
+  final String iconPath;
+  final VoidCallback? onTap;
+
+  const CategoryItem({
+    super.key,
+    required this.categoryName,
+    required this.iconPath,
+    this.onTap,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final colorScheme = Theme.of(context).colorScheme;
+
+    return InkWell(
+      onTap: onTap,
+      borderRadius: AppBorderRadius.large,
+      child: Container(
+        height: 100,
+        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+        decoration: BoxDecoration(
+          borderRadius: AppBorderRadius.large,
+          color: colorScheme.secondaryContainer.withAlpha(100),
+        ),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.start,
+              children: [
+                SvgPicture.asset(
+                  iconPath,
+                  width: 30,
+                  height: 30,
+                  colorFilter: ColorFilter.mode(
+                    colorScheme.primary,
+                    BlendMode.srcIn,
+                  ),
+                ),
+              ],
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [
+                Text(
+                  categoryName,
+                  style: TextStyle(
+                    color: colorScheme.primary,
+                    fontSize: 18,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
